requirements.txt
boto3>=1.26
python-dotenv>=1.0
demucs>=4.0         # <-- installs the CLI